refactor(memory_server): app/memory_server.py 转包 + 抽无状态叶子 rows/gates（阶段一）
按 game_router runtime 模式拆 4576 行上帝文件的第一刀（纯搬移零行为变化）：

- app/memory_server.py → app/memory_server/__init__.py（git mv，门面 +
  暂存其余全部符号；_repo_root 多爬一级目录）
- _shared.py：日志初始化（依赖根，不 import 任何兄弟模块）
- rows.py：纯行/消息提取函数（_coerce_db_ts、_extract_*_from_rows、
  _trim_to_user_msg_bracket、_has_human_messages、_extract_ai_response、
  _extract_user_messages）
- gates.py：维护状态（_maint_state 族）、功能开关（_ais_review_enabled /
  _ais_powerful_memory_enabled）、空闲判定（_touch_activity/_is_idle）、
  循环调度常量（IDLE_*/REVIEW_*/_INITIAL_DELAY_*）
- __main__.py：原 `if __name__ == "__main__"` 块，直跑改
  `python -m app.memory_server`（CONTRIBUTING.md 同步）

消费侧规则：gates 的可重绑状态与被测试 patch 的函数一律走
`gates.<attr>` 模块属性访问（晚绑定）；rows 纯函数 from-import。
门面 re-export 旧模块全部顶层符号，唯二例外 _maint_state /
_last_activity_time（会被原地重绑，快照会静默过期，docstring 已注明）。

测试 patch 目标随消费点迁移：test_memory_review_backoff /
test_recent_compress_backup / test_memory_server_cache_settle 改打
memory_server.gates。守卫脚本路径 app/memory_server.py →
app/memory_server（trailing_slash / async_blocking / no_temperature）。
