+++ CONTRIBUTING.md
@@ -44,7 +44,7 @@ cd N.E.K.O
 uv sync
 
 # Start services
-uv run python app/memory_server.py
+uv run python -m app.memory_server
 uv run python app/main_server.py
 # Optional: uv run python app/agent_server.py
 ```
@@ -126,7 +126,7 @@ cd N.E.K.O
 uv sync
 
 # 启动服务
-uv run python app/memory_server.py
+uv run python -m app.memory_server
 uv run python app/main_server.py
 # 可选：uv run python app/agent_server.py
 ```
