+++ scripts/check_api_trailing_slash.py
@@ -100,7 +100,7 @@ REPO_ROOT = Path(__file__).resolve().parent.parent
 DEFAULT_PATHS: list[str] = [
     "main_routers",
     "app/main_server.py",
-    "app/memory_server.py",
+    "app/memory_server",
     "app/agent_server.py",
     "plugin/server/routes",
 ]
