+++ scripts/check_async_blocking.py
@@ -85,8 +85,8 @@ DEFAULT_PATHS = [
     "main_logic",
     "main_routers",
     "utils",
-    # memory_server.py + its import chain (memory/ package)
-    "app/memory_server.py",
+    # memory_server package + its import chain (memory/ package)
+    "app/memory_server",
     "memory",
     # agent_server.py + its import chain (brain/ package; main_logic already
     # in scope above)
